requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-model-utils==1.1.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.1
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.3
